chore: rename and move filter options component
